config-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.25.0
python-dotenv>=1.0.0
Pillow>=9.0.0
colorgram.py>=1.2.0 